Fix cmd_line input, divide into files
